+++ HISTORY.md
@@ -41,7 +41,6 @@
     * TuneScope extension update, thanks, Harsh, Jo and Glen!
     * SciSnap2 extension update, thanks, Eckart!
     * MQTT library update, thanks, Simon!
-    * optimized canvas for reading back image data (slows down rendering, turned off by default)
     * removed PIPE from all libraries (b/c it is now a built-in primitive in the control blocks palette)
     * updated "Codification" example project
     * hide "stage selected - no motion primitives" text when hiding the category names in single palette mode
@@ -97,6 +96,8 @@
 ### 2023-10-31
 * blocks, threads, objects, lists: alphanumeric indices for lists, experimental
 * byob: fixed "unused blocks" dialog
+* morphic, gui: removed experimental canvas image data optimization
+* v8.1-rc2
 
 ### 2023-10-30
 * gui: added spriteNamed() helper